2006-06-17  Chris Toshok  <toshok@ximian.com>
	* SCSharp/SCSharp.Mpq/Dat.cs: new superclass for all *.dat files.
	Makes managing the variable blocks a little easier.  probably a
	little less performant, but I think overall it'll be a win.

	* SCSharp/SCSharp.Mpq/*Dat.cs: port to the new Dat.cs superclass.

	* SCSharp/SCSharp.UI/*.cs: same.  Also, clean up some compiler
	warnings and (CursorAnimator.cs) speed up the cursor animation.
	


git-svn-id: svn+ssh://scsharp.hungry.com/svn/scsharp/trunk@400 ea3e6beb-b7ab-da11-bdaa-00e0814018bc

+++ SCSharp/SCSharp.Mpq/IScriptBin.cs
@@ -39,13 +39,13 @@ namespace SCSharp
 	public class IScriptBin : MpqResource
 	{
 		byte[] buf;
-		Dictionary<ushort,ushort> entries;
+		Dictionary<uint,ushort> entries;
 
 		const int entry_table_offset = 0x0082e0;
 
 		public IScriptBin ()
 		{
-			entries = new Dictionary<ushort,ushort>();
+			entries = new Dictionary<uint,ushort>();
 		}
 
 		public void ReadFromStream (Stream stream)
@@ -71,7 +71,7 @@ namespace SCSharp
 			get { return buf; }
 		}
 
-		public ushort GetScriptEntryOffset (ushort images_id) {
+		public ushort GetScriptEntryOffset (uint images_id) {
 			if (!entries.ContainsKey (images_id))
 				return 0;
 			return entries[images_id];
